new assignment

+++ src/WeeklyAssignment/WeeklyAssignments_Progarm1.java
@@ -7,34 +7,34 @@ public class WeeklyAssignments_Progarm1 {
 
 		// Problem Statement:1
 		float marks = 87.6f;
-		//upcasting   
+		//upcasting
 		double dmarks = marks;
 		//downcasting
 		int imarks = (int)marks;
 		//(ASCII conversion)
 		char cmarks= (char)marks;
-		
+
 
 		System.out.println("Float value: "+ marks);
-		System.out.println("Double value: "+ dmarks);	
+		System.out.println("Double value: "+ dmarks);
 		System.out.println("Integer value: "+ imarks);
 	    System.out.println("Character value: "+ cmarks);
 	    System.out.println("ðŸ‘‰ ASCII 97 = "+ "'" + cmarks+ "'");
-	    
+
 	    //// Problem Statement:2 ASCII Explorer
-	    
-	  
+
+
 	    char input ='A';
-	    int value2 = (int)input;
+	    int value2 = input;
 	    System.out.println("ASCII value "+ value2);
 	    int newASCII = input+5;
 	    char newChar = (char)newASCII;
-	    
+
 	    System.out.println("New ASCII value: "+ newASCII );
 	    System.out.println("New character: "+ newChar );
-	    
-	    
-		
+
+
+
 	}
 
 }
